Simplify plugin files structure
Plugin methods no longer live in a subpackage.
