+++ logger.go
@@ -1,0 +1,50 @@
+/*
+http://www.apache.org/licenses/LICENSE-2.0.txt
+
+
+Copyright 2016 Intel Corporation
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package main
+
+import (
+	"strings"
+
+	log "github.com/Sirupsen/logrus"
+	"github.com/intelsdi-x/snap-plugin-lib-go/v1/plugin"
+)
+
+func getLogger(cfg plugin.Config) *log.Entry {
+	logger := log.WithFields(log.Fields{
+		"plugin-name":    Name,
+		"plugin-version": Version,
+		"plugin-type":    "publisher",
+	})
+
+	log.SetLevel(log.WarnLevel)
+
+	levelValue, err := cfg.GetString("log-level")
+	if err == nil {
+		if level, err := log.ParseLevel(strings.ToLower(levelValue)); err == nil {
+			log.SetLevel(level)
+		} else {
+			log.WithFields(log.Fields{
+				"value":             strings.ToLower(levelValue),
+				"acceptable values": "warn, error, debug, info",
+			}).Warn("Invalid config value")
+		}
+	}
+	return logger
+}
