+++ publisher_test.go
@@ -19,7 +19,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package graphite
+package main
 
 import (
 	"testing"
